requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
plotly>=5.18
folium>=0.15
streamlit-folium>=0.18
Pillow>=10.0
earthengine-api>=0.1.385
torch>=2.1
matplotlib>=3.8
flask>=3.0
requests>=2.31
pyserial>=3.5
gunicorn>=21.2
